refactor: share headed evidence environment
