+++ scripts/final-cut-canonical-headed-e2e.mjs
@@ -1,16 +1,11 @@
 import { createHash } from "node:crypto";
-import { readFile } from "node:fs/promises";
-import os from "node:os";
-import { execFile as execFileCallback } from "node:child_process";
 import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
 import { fileURLToPath } from "node:url";
 import { dirname, join } from "node:path";
-import { promisify } from "node:util";
-import { sanitizeCanonicalEvidence } from "./final-cut-evidence.mjs";
+import { evidenceEnvironment, sanitizeCanonicalEvidence } from "./final-cut-evidence.mjs";
 
 const root = join(dirname(fileURLToPath(import.meta.url)), "..");
-const execFile = promisify(execFileCallback);
 const expectedProject = process.env.FRAMEKIT_FINAL_CUT_E2E_PROJECT;
 const clipId = process.env.FRAMEKIT_FINAL_CUT_E2E_CLIP_ID;
 
@@ -83,7 +78,7 @@ try {
     diff: transaction.diff,
     restored,
     digests: { before: beforeDigest, restored: restoredDigest },
-  }, await evidenceEnvironment());
+  }, await evidenceEnvironment(root));
   process.stdout.write(`${JSON.stringify(evidence, null, 2)}\n`);
 } catch (error) {
   if (transactionId) {
@@ -130,34 +125,4 @@ function stableJson(value) {
 
 function compareText(left, right) {
   return left < right ? -1 : left > right ? 1 : 0;
-}
-
-async function evidenceEnvironment() {
-  const packageJson = JSON.parse(await readFile(join(root, "package.json"), "utf8"));
-  let gitCommit;
-  try {
-    gitCommit = (await execFile("git", ["rev-parse", "HEAD"], { cwd: root })).stdout.trim();
-  } catch (error) {
-    throw new Error(`FINAL_CUT_E2E_COMMIT_UNAVAILABLE: ${String(error)}`);
-  }
-  if (!gitCommit) throw new Error("FINAL_CUT_E2E_COMMIT_UNAVAILABLE: git returned an empty commit");
-  return {
-    framekitVersion: packageJson.version,
-    finalCutVersion: await finalCutVersion(),
-    gitCommit,
-    nodeVersion: process.version,
-    platform: process.platform,
-    architecture: process.arch,
-    osVersion: os.version(),
-  };
-}
-
-async function finalCutVersion() {
-  try {
-    const version = (await execFile("osascript", ["-e", 'tell application "Final Cut Pro" to get version'])).stdout.trim();
-    if (version) return version;
-  } catch (error) {
-    throw new Error(`FINAL_CUT_E2E_FINAL_CUT_VERSION_UNAVAILABLE: ${String(error)}`);
-  }
-  throw new Error("FINAL_CUT_E2E_FINAL_CUT_VERSION_UNAVAILABLE: Final Cut Pro returned an empty version");
 }
