+++ scripts/final-cut-canonical-read-headed-e2e.mjs
@@ -1,15 +1,10 @@
-import { readFile } from "node:fs/promises";
-import os from "node:os";
-import { execFile as execFileCallback } from "node:child_process";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
-import { promisify } from "node:util";
 import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
-import { sanitizeCanonicalReadEvidence } from "./final-cut-evidence.mjs";
+import { evidenceEnvironment, sanitizeCanonicalReadEvidence } from "./final-cut-evidence.mjs";
 
 const root = join(dirname(fileURLToPath(import.meta.url)), "..");
-const execFile = promisify(execFileCallback);
 const expectedProject = process.env.FRAMEKIT_FINAL_CUT_E2E_PROJECT;
 const expectedSequence = process.env.FRAMEKIT_FINAL_CUT_E2E_SEQUENCE_ID;
 
@@ -74,7 +69,7 @@ try {
     project: { id: project.id, name: project.name, sequenceId: sequence.id },
     catalog,
     snapshot,
-  }, await evidenceEnvironment());
+  }, await evidenceEnvironment(root));
   process.stdout.write(`${JSON.stringify(evidence, null, 2)}\n`);
 } finally {
   await client.close().catch(() => {});
@@ -90,21 +85,4 @@ async function callJson(name, arguments_ = {}) {
   } catch {
     throw new Error(`${name} returned invalid JSON: ${text}`);
   }
-}
-
-async function evidenceEnvironment() {
-  const packageJson = JSON.parse(await readFile(join(root, "package.json"), "utf8"));
-  const gitCommit = (await execFile("git", ["rev-parse", "HEAD"], { cwd: root })).stdout.trim();
-  if (!/^[0-9a-f]{40}$/i.test(gitCommit)) throw new Error("FINAL_CUT_E2E_COMMIT_UNAVAILABLE: git returned an invalid commit");
-  const finalCutVersion = (await execFile("osascript", ["-e", 'tell application "Final Cut Pro" to get version'])).stdout.trim();
-  if (!finalCutVersion) throw new Error("FINAL_CUT_E2E_FINAL_CUT_VERSION_UNAVAILABLE: Final Cut Pro returned an empty version");
-  return {
-    framekitVersion: packageJson.version,
-    finalCutVersion,
-    gitCommit,
-    nodeVersion: process.version,
-    platform: process.platform,
-    architecture: process.arch,
-    osVersion: os.version(),
-  };
 }
