+++ scripts/final-cut-evidence.d.mts
@@ -1,3 +1,13 @@
+export function evidenceEnvironment(root: string): Promise<{
+  framekitVersion: string;
+  finalCutVersion: string;
+  gitCommit: string;
+  nodeVersion: string;
+  platform: string;
+  architecture: string;
+  osVersion: string;
+}>;
+
 export function sanitizeCanonicalEvidence(run: unknown, environment: {
   framekitVersion: string;
   finalCutVersion: string;
